requirements.txt
# Core dependencies
langchain>=0.1.5
langchain_community>=0.0.15
langchain_huggingface>=0.0.1
langchain_chroma>=0.0.10
langchain_openai>=0.0.5
langchain_groq>=0.0.2
langchain_text_splitters>=0.0.1
langchain_core>=0.1.7

# Vector database
chromadb>=0.4.22

# User interface
gradio>=4.12.0

# API and data handling
youtube-transcript-api>=0.6.1
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2

# AI models
transformers>=4.37.2
sentence-transformers>=2.2.2
groq>=0.4.0
openai>=1.5.0

# NLP tools
nltk>=3.8.1

# Data processing
tqdm>=4.66.1
numpy>=1.24.0
pandas>=2.1.0
pydantic>=1.10.8

# Utilities
ipython>=8.18.0